requirements.txt
langgraph>=0.2.0
langchain>=0.2.0
langchain-openai>=0.1.0
langchain-anthropic>=0.1.0
langchain_community>=0.2.0
pydantic>=2.0.0
rich>=13.0.0
python-dotenv>=1.0.0
httpx>=0.28.1
markdownify>=1.2.0
